refactor: Label 컴포넌트 onClick Prop 추가 및 shared 레이어로 위치 변경
