+++ src/shared/ui/label/Label.tsx
@@ -7,6 +7,7 @@ type LabelProps = {
   borderColor?: string;
   width: number;
   height: number;
+  onClick?: () => void;
 };
 
 export default function Label({
@@ -16,6 +17,7 @@ export default function Label({
   backgroundColor,
   width,
   height,
+  onClick,
 }: LabelProps) {
   return (
     <S.LabelContainer
@@ -24,6 +26,7 @@ export default function Label({
       radius="24px"
       width={width}
       height={height}
+      onClick={onClick}
     >
       <S.LabelText $fontColor={fontColor}>{text}</S.LabelText>
     </S.LabelContainer>
